Wrote service in test file by accident, fixing
